requirements.txt
mysqlclient~=2.1.1
requests~=2.28.2
Flask~=1.1.4
numpy~=1.24.2
markupsafe==2.0.1
google-api-python-client==2.81.0
google-auth==2.16.2
SQLAlchemy==1.3
cloud-sql-python-connector~=1.3.0
pymysql==1.1.0